shift module by complexity
