+++ server/src/modules/module-6/exercise/package.json
@@ -1,24 +1,28 @@
 {
-  "name": "module-6-middleware-exercise",
+  "name": "module-6-pagination-exercise",
   "version": "1.0.0",
-  "description": "Route protection middleware exercise for Express.js middleware",
+  "description": "Exercise for learning pagination in Express.js",
   "main": "server.js",
   "scripts": {
     "start": "node server.js",
-    "test": "mocha test.js"
-  },
-  "keywords": ["express", "middleware", "authentication", "api", "exercise"],
-  "author": "Backend Playground",
-  "license": "MIT",
-  "engines": {
-    "node": ">=14.0.0"
+    "test": "mocha test.js --timeout 10000",
+    "solution": "node solution.js"
   },
   "dependencies": {
-    "express": "^4.18.0"
+    "express": "^4.18.2"
   },
   "devDependencies": {
     "chai": "^4.3.7",
     "mocha": "^10.2.0",
-    "supertest": "^6.3.0"
-  }
+    "supertest": "^6.3.3"
+  },
+  "keywords": [
+    "express",
+    "pagination",
+    "api",
+    "nodejs",
+    "backend"
+  ],
+  "author": "Backend Playground",
+  "license": "MIT"
 }
