+++ server/src/modules/module-8/exercise/package.json
@@ -1,33 +1,21 @@
 {
-  "name": "module-8-authentication",
+  "name": "module-8-data-persistence",
   "version": "1.0.0",
-  "description": "Authentication & Authorization Exercise",
+  "description": "Module 8 Exercise: User Data Storage API",
   "main": "server.js",
   "scripts": {
-    "start": "node server.js",
-    "dev": "nodemon server.js",
-    "test": "mocha test.js --timeout 10000"
+    "test": "mocha test.js",
+    "start": "node server.js"
   },
+  "keywords": ["data-persistence", "file-system", "json", "nodejs", "express", "api"],
+  "author": "Backend Playground",
+  "license": "MIT",
   "dependencies": {
-    "express": "^4.18.2",
-    "bcryptjs": "^2.4.3",
-    "jsonwebtoken": "^9.0.2"
+    "express": "^4.18.0"
   },
   "devDependencies": {
-    "chai": "^4.3.7",
-    "mocha": "^10.2.0",
-    "supertest": "^6.3.3",
-    "nodemon": "^3.0.1"
-  },
-  "keywords": [
-    "authentication",
-    "authorization",
-    "jwt",
-    "bcrypt",
-    "express",
-    "api",
-    "security"
-  ],
-  "author": "Backend Playground",
-  "license": "MIT"
+    "mocha": "^10.0.0",
+    "chai": "^4.3.0",
+    "supertest": "^6.3.0"
+  }
 }
