+++ server/src/modules/module-9/exercise/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "module-9-pagination-exercise",
+  "name": "module-9-authentication-exercise",
   "version": "1.0.0",
-  "description": "Exercise for learning pagination in Express.js",
+  "description": "Exercise for learning authentication and security in Express.js",
   "main": "server.js",
   "scripts": {
     "start": "node server.js",
@@ -9,7 +9,9 @@
     "solution": "node solution.js"
   },
   "dependencies": {
-    "express": "^4.18.2"
+    "express": "^4.18.2",
+    "bcrypt": "^5.1.0",
+    "jsonwebtoken": "^9.0.0"
   },
   "devDependencies": {
     "chai": "^4.3.7",
